Separate dapui and goide functionality to their own files
